--- src/Sylius/Bundle/CoreBundle/Form/EventSubscriber/AddAuthorGuestTypeFormSubscriber.php
@@ -1,54 +1,0 @@
-<?php
-
-/*
- * This file is part of the Sylius package.
- *
- * (c) Paweł Jędrzejewski
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
-namespace Sylius\Bundle\CoreBundle\Form\EventSubscriber;
-
-use Sylius\Component\Review\Model\ReviewInterface;
-use Symfony\Component\EventDispatcher\EventSubscriberInterface;
-use Symfony\Component\Form\FormEvent;
-use Symfony\Component\Form\FormEvents;
-use Symfony\Component\Validator\Constraints\Valid;
-use Webmozart\Assert\Assert;
-
-/**
- * @internal
- *
- * @author Arkadiusz Krakowiak <arkadiusz.krakowiak@lakion.com>
- */
-final class AddAuthorGuestTypeFormSubscriber implements EventSubscriberInterface
-{
-    /**
-     * {@inheritdoc}
-     */
-    public static function getSubscribedEvents()
-    {
-        return [
-            FormEvents::PRE_SET_DATA => 'preSetData',
-        ];
-    }
-
-    /**
-     * @param FormEvent $event
-     */
-    public function preSetData(FormEvent $event)
-    {
-        $form = $event->getForm();
-        /** @var ReviewInterface $review */
-        $review = $event->getData();
-        $author = $form->getConfig()->getOption('author');
-
-        Assert::isInstanceOf($review, ReviewInterface::class);
-
-        if (null === $author && null === $review->getAuthor()) {
-            $form->add('author', 'sylius_customer_guest', ['constraints' => [new Valid()]]);
-        }
-    }
-}
