+++ src/Sylius/Bundle/CoreBundle/Form/Type/Product/ProductReviewType.php
@@ -11,15 +11,18 @@
 
 namespace Sylius\Bundle\CoreBundle\Form\Type\Product;
 
-use Sylius\Bundle\CoreBundle\Form\EventSubscriber\AddAuthorGuestTypeFormSubscriber;
 use Sylius\Bundle\ReviewBundle\Form\Type\ReviewType;
+use Sylius\Component\Review\Model\ReviewInterface;
 use Symfony\Component\Form\FormBuilderInterface;
-use Symfony\Component\OptionsResolver\OptionsResolver;
+use Symfony\Component\Form\FormEvent;
+use Symfony\Component\Form\FormEvents;
+use Symfony\Component\Validator\Constraints\Valid;
+use Webmozart\Assert\Assert;
 
 /**
  * @author Arkadiusz Krakowiak <arkadiusz.krakowiak@lakion.com>
  */
-class ProductReviewType extends ReviewType
+final class ProductReviewType extends ReviewType
 {
     /**
      * {@inheritdoc}
@@ -28,18 +31,31 @@ class ProductReviewType extends ReviewType
     {
         parent::buildForm($builder, $options);
 
-        $builder->addEventSubscriber(new AddAuthorGuestTypeFormSubscriber());
+        $builder->addEventListener(FormEvents::PRE_SET_DATA, function (FormEvent $event) use ($options) {
+            $form = $event->getForm();
+            $review = $event->getData();
+
+            Assert::isInstanceOf($review, ReviewInterface::class);
+
+            if (null === $review->getAuthor()) {
+                $form->add('author', 'sylius_customer_guest', ['constraints' => [new Valid()]]);
+            }
+        });
     }
 
     /**
      * {@inheritdoc}
      */
-    public function configureOptions(OptionsResolver $resolver)
+    public function getName()
     {
-        parent::configureOptions($resolver);
+        return 'sylius_product_review';
+    }
 
-        $resolver->setDefaults([
-            'author' => null,
-        ]);
+    /**
+     * {@inheritdoc}
+     */
+    public function getBlockPrefix()
+    {
+        return 'sylius_product_review';
     }
 }
