+++ src/Sylius/Bundle/CoreBundle/Resources/config/app/sylius/sylius_review.yml
@@ -9,8 +9,7 @@ sylius_review:
                 classes:
                     model: Sylius\Component\Core\Model\ProductReview
                     repository: Sylius\Bundle\CoreBundle\Doctrine\ORM\ProductReviewRepository
-                    form:
-                        default: Sylius\Bundle\CoreBundle\Form\Type\Product\ProductReviewType
+                    form: Sylius\Bundle\CoreBundle\Form\Type\Product\ProductReviewType
             reviewer:
                 classes:
                     model: "%sylius.model.customer.class%"
