+++ src/Sylius/Bundle/CoreBundle/Resources/config/services/form.xml
@@ -65,6 +65,12 @@
             <tag name="form.type_extension" extended-type="Sylius\Bundle\TaxonomyBundle\Form\Type\TaxonType" />
         </service>
 
+        <service id="sylius.form.type.product_review" class="Sylius\Bundle\CoreBundle\Form\Type\Product\ProductReviewType">
+            <argument>%sylius.model.product_review.class%</argument>
+            <argument type="collection"><argument>sylius</argument></argument>
+            <tag name="form.type" />
+        </service>
+
         <service id="sylius.form.type.product_image" class="%sylius.form.type.product_image.class%">
             <argument>%sylius.model.product_image.class%</argument>
             <tag name="form.type" alias="sylius_product_image" />
