+++ src/Sylius/Bundle/ShopBundle/Resources/config/routing/product_review.yml
@@ -20,7 +20,6 @@ sylius_shop_product_reviews_create:
         _sylius:
             template: SyliusShopBundle:Product/Review:create.html.twig
             form:
-                type: sylius_product_review
                 options:
                     validation_groups: ['sylius', 'sylius_review']
             factory:
